@lhotari 2024-09-11
This introduces a regression that is fixed by #23284